+++ fabric/src/main/java/com/unrealdinnerbone/jamd/data/JAMDData.java
@@ -36,6 +36,7 @@ public class JAMDData implements DataGeneratorEntrypoint {
         pack.addProvider(LootTableProvider::new);
         pack.addProvider(AdvancementProvider::new);
         pack.addProvider(FeatureTagProvider::new);
+        pack.addProvider(PlaceTagProvider::new);
     }
 
     @Override