+++ fabric/src/main/java/com/unrealdinnerbone/jamd/data/PlaceTagProvider.java
@@ -1,0 +1,26 @@
+package com.unrealdinnerbone.jamd.data;
+
+import com.unrealdinnerbone.jamd.JAMDRegistry;
+import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricTagProvider;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.data.worldgen.features.OreFeatures;
+import net.minecraft.data.worldgen.placement.OrePlacements;
+import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
+import net.minecraft.world.level.levelgen.placement.PlacedFeature;
+
+import java.util.concurrent.CompletableFuture;
+
+public class PlaceTagProvider extends FabricTagProvider<PlacedFeature> {
+    public PlaceTagProvider(FabricDataOutput output, CompletableFuture<HolderLookup.Provider> registriesFuture) {
+        super(output, Registries.PLACED_FEATURE, registriesFuture);
+    }
+
+    @Override
+    protected void addTags(HolderLookup.Provider arg) {
+//        getOrCreateTagBuilder(JAMDRegistry.OVERWORLD.getIgnoredPlaceFeatures())
+//                .add(OrePlacements.ORE_CLAY);
+
+    }
+}